resolves #275 : edits to a_transportation_problem.rst (#277)
* Update a_transportation_problem.rst to reference BeerDistributionProblem.py

This fixes code that was missing in the example docs to make the PuLP model for the transportation case study.

* Update a_transportaion_problem.rst with link

Add link to the example .py file

* Create BeerDistributionProblemWarehouseExtension.py

This file contains the code for solving the beer distribution case study with an extra warehouse

* Create BeerDistributionProblemCompetitorExtension.py

This file contains the code for the Beer Distribution Problem when beer from a competitor is added

* Update extension problems to reference new files in /examples

Updated a_transportation_problem.rst to reference corresponding (new) files for the extensions of adding an extra warehouse and a competitor. The competitor code was also changed to better reflect the graphics in the case study. Each extension sub-problem also contains a link to the .py file