Update wedding.py (#323)
* Update wedding.py

Suggestion to use lpSum instead of sum to make code more efficient and aligned with best practice and the other cases. See https://stackoverflow.com/questions/63081381/when-to-use-sum-v-s-lpsum-using-pulp/63081480#63081480 for background. 

The code runs significantly faster.

Note that the import statement varies among the different pulp cases. In this example and the "A Two Stage Production Planning Problem" the statement used is "import pulp". 

Also note that this is my first PR.

* Update wedding_initial.py

Use pulp.lpSum instead of Sum

* Update how_to_export_models.rst

Use pulp.lpSum instead of sum.