IPython printing fix for CBC+GUROBI solvers api (coin-or#333)
If I try to print log with GUROBI_CMD(msg=1) function it will not work when using an IPython environment like Jupyter Notebook or Spyder. It works perfectly if the script is run from the terminal but I need Spyder cells to teach. I can use %%python cell magic but then I would have to run all the script of an optimization model in a cell, which is not the best workaround when you need to teach with the least cognitive load as possible (not opening the console in the middle of the class). I only changed the default pipe: None for subprocess.PIPE to capture the terminal output and then I print it in real time. I am of course not the best programmer but I hope this helps. I leave the compatibility to you, thanks!!! 😊✨

IPython printing fix for CBC+GUROBI solvers api (coin-or#333)

If I try to print log with GUROBI_CMD(msg=1) function it will not work when using an IPython environment like Jupyter Notebook or Spyder. It works perfectly if the script is run from the terminal but I need Spyder cells to teach. I can use %%python cell magic but then I would have to run all the script of an optimization model in a cell, which is not the best workaround when you need to teach with the least cognitive load as possible (not opening the console in the middle of the class). I only changed the default pipe: None for subprocess.PIPE to capture the terminal output and then I print it in real time. I am of course not the best programmer but I hope this helps. I leave the compatibility to you, thanks!!! 😊✨