Xpress api improvements (#557)
* Explicitly create a file with command input.

This avoids potential blocking on pipe writing in case the number of
input commands becomes huge.

* Fix detection of infeasibility.

Instead of guessing infeasibility from the violation, we now explicitly
capture the status reported by Xpress. From that we can easily derive
the optimization status and map to the respective PuLP constants.

* Support warm starting the solver.

* Fix link to optimizer control reference.

We should point to the official Xpress documentation.

* Extract row and dual information from PRT file.

* Suppress Xpress banner if messages are disabled.

* Apply formatting.

* Explain why we don't use os.devnull.

* Fix handling of infeasible problems.

The PRT file may be short or empty in case the problem was found
infeasible.

* Fix expected test results for the Xpress solver.

The solver will raise an error in case names are too long for the LP
format. This is because the solver goes through the command line and
thus has to communicate via LP files.

* Fix expected test results for the Xpress solver.

The solver will raise an error in case of duplicate name.
This is because the solver goes through the command line and
thus has to communicate via LP files.

* Don't use shell's `echo`, instead use TCL builtin stuff.

* Mention warmstart support in documentation.

* Fix checking for presence of arguments.

* Xpress: quote paths properly for Tcl shell

Co-authored-by: Chris Brown <cbrown@fico.com>