CI: Update docs and publish workflows
A bit of maintenance on the docs and publish GitHub Actions workflows.

Both are updated to the latest versions of checkout and setup-python, as wel as a recent Python version and the latest Ubuntu image.

Using the master branch of pypa/gh-action-pypi-publish is deprecated, so the release/v1 branch is now used.